Complete installation workflow for uploaded file

+++ ModuleInstaller.php
@@ -1,7 +1,11 @@
 <?php
 
 namespace ICM\ModuleInstaller;
-include 'utils/functions.php';
+
+include_once 'utils/functions.php';
+include_once 'utils/ModuleInstallerException.php';
+
+use ICM\ModuleInstaller\ModuleInstallerException;
 
 define('APP_PATH_EXTERNAL_MODULES', dirname(APP_PATH_DOCROOT) . DIRECTORY_SEPARATOR . 'modules');
 define('TEMP_FOLDER', APP_PATH_TEMP . DIRECTORY_SEPARATOR . 'module_installer');
@@ -22,12 +26,12 @@ class ModuleInstaller extends \ExternalModules\AbstractExternalModule
         parent::__construct();
 
         if (file_exists(TEMP_FOLDER) && !is_dir(TEMP_FOLDER)) {
-            throw new \Exception("The temporary folder is not a directory.");
+            throw new ModuleInstallerException("The temporary folder is not a directory.");
         }
         if (!file_exists(TEMP_FOLDER)) {
             // Create the temporary folder
             if (!mkdir(TEMP_FOLDER, 0777, true)) {
-                throw new \Exception("The temporary folder could not be created.");
+                throw new ModuleInstallerException("The temporary folder could not be created.");
             }
         }
     }
@@ -37,7 +41,7 @@ class ModuleInstaller extends \ExternalModules\AbstractExternalModule
      *
      * @param $uploaded_file
      * @return string
-     * @throws \Exception
+     * @throws ModuleInstallerException
      */
     public function moveZipFileToTempLocation($filename): string
     {
@@ -46,7 +50,7 @@ class ModuleInstaller extends \ExternalModules\AbstractExternalModule
 
         // Move the uploaded file to the temporary location
         if (!move_uploaded_file($filename, TEMP_FOLDER . '/' . $temp_filename)) {
-            throw new \Exception("The uploaded file could not be moved to a temporary location.");
+            throw new ModuleInstallerException("The uploaded file could not be moved to a temporary location.");
         }
         return $temp_filename;
     }
@@ -75,19 +79,19 @@ class ModuleInstaller extends \ExternalModules\AbstractExternalModule
                 $zip->extractTo(TEMP_FOLDER);
                 $zip->close();
             } else {
-                throw new \Exception("The ZIP file could not be opened.");
+                throw new ModuleInstallerException("The ZIP file could not be opened.");
             }
 
             // Check that the module contains a config.json file and read it
             if (!file_exists(TEMP_FOLDER . '/' . $module_info->parent_folder_name . '/config.json')) {
-                throw new \Exception("The ZIP file does not contain a config.json file.");
+                throw new ModuleInstallerException("The ZIP file does not contain a config.json file.");
             }
 
             // Get the module name from the config.json file
             $config = json_decode(file_get_contents(TEMP_FOLDER . '/' . $module_info->parent_folder_name . '/config.json'), true);
 
             if (!isset($config['name'])) {
-                throw new \Exception("The config.json file does not contain a name.");
+                throw new ModuleInstallerException("The config.json file does not contain a name.");
             }
             $module_info->name = $config['name'];
 
@@ -103,10 +107,12 @@ class ModuleInstaller extends \ExternalModules\AbstractExternalModule
             $module_info->path = strtolower(preg_replace('/[^a-zA-Z0-9]/', '_', $path_from_name));
 
         }
-        catch (\Exception $e) {
-            throw new \Exception($e->getMessage());
+        catch (ModuleInstallerException $e) {
+            throw new ModuleInstallerException($e->getMessage());
         } finally {
-            delete_directory(TEMP_FOLDER . '/' . $module_info->parent_folder_name);
+            if (is_dir(TEMP_FOLDER . '/' . $module_info->parent_folder_name)) {
+                delete_directory(TEMP_FOLDER . '/' . $module_info->parent_folder_name);
+            }
         }
 
         return $module_info;
@@ -119,14 +125,14 @@ class ModuleInstaller extends \ExternalModules\AbstractExternalModule
         // TODO: allow? With a warning?
         // Check if the module already exists
         if (file_exists(APP_PATH_EXTERNAL_MODULES . '/' . $module_path)) {
-            throw new \Exception("The module already exists in $module_path.");
+            throw new ModuleInstallerException("The module already exists in $module_path.");
         }
 
         try {
             // Rename the parent folder of the zip to the module path
             $zip = new \ZipArchive;
             if ($zip->open(TEMP_FOLDER . DIRECTORY_SEPARATOR . $tmp_filename) !== TRUE) {
-                throw new \Exception("The ZIP file could not be opened.");
+                throw new ModuleInstallerException("The ZIP file could not be opened.");
             }
             $i = 0;
             while ($item_name = $zip->getNameIndex($i)){
@@ -141,12 +147,36 @@ class ModuleInstaller extends \ExternalModules\AbstractExternalModule
                 $zip->extractTo(APP_PATH_EXTERNAL_MODULES);
                 $zip->close();
             } else {
-                throw new \Exception("The ZIP file could not be opened.");
+                throw new ModuleInstallerException("The ZIP file could not be opened.");
             }
         }
         catch (\Exception $e) {
-            throw new \Exception($e->getMessage());
+            throw new ModuleInstallerException($e->getMessage());
         }
     }
 
+    /**
+     * Returns the list of installed versions of a module.
+     *
+     * @param $module_name
+     * @return array
+     */
+    public function installedVersionsOf($module_name): array
+    {
+        // Loop through the folders into the modules folder
+        $installed_versions = [];
+        foreach (scandir(APP_PATH_EXTERNAL_MODULES) as $item) {
+            if ($item == '.' || $item == '..') {
+                continue;
+            }
+            // Check if the folder contains a config.json file and read it
+            if (file_exists(APP_PATH_EXTERNAL_MODULES . '/' . $item . '/config.json')) {
+                $config = json_decode(file_get_contents(APP_PATH_EXTERNAL_MODULES . '/' . $item . '/config.json'), true);
+                if (isset($config['name']) && $config['name'] == $module_name) {
+                    $installed_versions[] = $item;
+                }
+            }
+        }
+        return $installed_versions;
+    }
 }
